make components for meta
