+++ packages/mars-theme/src/components/components/breadcrumbs/index.js
@@ -2,46 +2,40 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect, decode } from 'frontity';
 import Link from '../link';
-import Item from './item';
 import Divider from './divider';
 import { isSearchLink } from '../../helpers/content';
 import CategoryItem from './category-item';
 
-import { Wrapper } from './styles';
+import Wrapper from './styles';
 
 const Breadcrumbs = ({ state }) => {
   const data = state.source.get(state.router.link);
-  if (state.router.link === '/') {
-    return null;
-  }
 
   return (
-    <Wrapper className="breadcrumbs-wrapper">
-      <Item clickable>
-        <Link link="/">Blog Home</Link>
-      </Item>
+    <Wrapper>
+      <Link link="/">Blog Home</Link>
       {isSearchLink(state.router.link) && (
         <>
           <Divider />
-          <Item>Search results</Item>
+          <span>Search results</span>
         </>
       )}
       {data.isAuthor && (
         <>
           <Divider />
-          <Item>{decode(state.source.author[data.id].name)}</Item>
+          <span>{decode(state.source.author[data.id].name)}</span>
         </>
       )}
       {data.isCategory && (
         <>
           <Divider />
-          <Item>{decode(state.source[data.taxonomy][data.id].name)}</Item>
+          <span>{decode(state.source[data.taxonomy][data.id].name)}</span>
         </>
       )}
       {data.isTag && (
         <>
           <Divider />
-          <Item>Tags</Item>
+          <span>Tags</span>
         </>
       )}
       {data.isPostType && (
@@ -49,7 +43,7 @@ const Breadcrumbs = ({ state }) => {
           <Divider />
           <CategoryItem />
           <Divider />
-          <Item>{decode(state.source[data.type][data.id].title.rendered)}</Item>
+          <span>{decode(state.source[data.type][data.id].title.rendered)}</span>
         </>
       )}
     </Wrapper>
