+++ packages/mars-theme/src/components/pages/error/index.js
@@ -6,7 +6,7 @@ const description404 = (
   <>
     That page can’t be found
     {' '}
-    <span role='img' aria-label='confused face'>
+    <span role="img" aria-label="confused face">
       😕
     </span>
   </>
@@ -15,7 +15,7 @@ const description404 = (
 const description = (
   <>
     Don&apos;t panic! Seems like you encountered an error. If this persists,
-    <a href='https://community.frontity.org'> let us know </a>
+    <a href="https://community.frontity.org"> let us know </a>
     {' '}
     or try refreshing
     your browser.
@@ -38,8 +38,8 @@ const Page404 = ({ state }) => {
 };
 
 Page404.propTypes = {
-  state: PropTypes.object
-}
+  state: PropTypes.object,
+};
 
 export default connect(Page404);
