+++ packages/mars-theme/src/components/pages/loading/index.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Loader } from 'gfw-components';
+
+import Container from './styles';
+
+const Loading = () => (
+  <Container>
+    <Loader />
+  </Container>
+);
+
+export default Loading;
